Move Trinkets-related classes into new compat package
